Force the button to be a button, rather than a submit.
Previously a submit action was being triggered for the button after the
jquery handler had fired. Neither e.preventDefault(), e.stopPropagation()
or e.immediateStopPropagation() changed this behaviour, apparently
because the button's submit event is not directly related to the
tap event.

By changing the button type to be explicitly 'button' no submit event is
generated and all is good.

Signed-off-by: Jon Seymour <jon.seymour@gmail.com>